split post tests into 2
